requirements.txt
# requirements
python==3.10.11
rasa==3.6.1
openai==0.27.8
python-dotenv==1.0.0
ujson
```

